+++ cmd/saucectl/saucectl.go
@@ -5,7 +5,7 @@ import (
 	"github.com/saucelabs/saucectl/cli/command/configure"
 	initCmd "github.com/saucelabs/saucectl/cli/command/init"
 	"github.com/saucelabs/saucectl/cli/command/run"
-	"github.com/saucelabs/saucectl/cli/command/signup"
+	"github.com/saucelabs/saucectl/internal/cmd/signup"
 	"github.com/saucelabs/saucectl/internal/setup"
 	"github.com/saucelabs/saucectl/internal/version"
 	"os"
